Correct errors installation instructions for Docker files (#156)
This pull request updates the installation instructions in the
documentation to reference the correct files for the public release and
environment setup. The changes clarify which files to download and where
to save them.

Documentation improvements:

* Updated the public release installation steps to instruct users to
download the correct `compose.yml` file and added instructions to
download the `nginx.conf` reverse proxy configuration.
* Corrected the link for downloading the sample environment file,
ensuring users save `.env.sample` as `.env` instead of downloading the
wrong file.

* Closes #155